very hacky but works for simple upload/download/preview, needs engineered/tested

+++ services/repository-managers/src/main/java/org/sagebionetworks/repo/manager/file/MultipartManagerV2Impl.java
@@ -1,5 +1,6 @@
 package org.sagebionetworks.repo.manager.file;
 
+import java.io.File;
 import java.io.UnsupportedEncodingException;
 import java.net.URL;
 import java.security.MessageDigest;
@@ -28,6 +29,8 @@ import org.sagebionetworks.repo.model.file.AddPartState;
 import org.sagebionetworks.repo.model.file.BatchPresignedUploadUrlRequest;
 import org.sagebionetworks.repo.model.file.BatchPresignedUploadUrlResponse;
 import org.sagebionetworks.repo.model.file.CompleteMultipartRequest;
+import org.sagebionetworks.repo.model.file.FileHandle;
+import org.sagebionetworks.repo.model.file.GoogleCloudFileHandle;
 import org.sagebionetworks.repo.model.file.MultipartUploadRequest;
 import org.sagebionetworks.repo.model.file.MultipartUploadState;
 import org.sagebionetworks.repo.model.file.MultipartUploadStatus;
@@ -35,12 +38,18 @@ import org.sagebionetworks.repo.model.file.PartMD5;
 import org.sagebionetworks.repo.model.file.PartPresignedUrl;
 import org.sagebionetworks.repo.model.file.PartUtils;
 import org.sagebionetworks.repo.model.file.S3FileHandle;
+import org.sagebionetworks.repo.model.file.S3FileHandleInterface;
+import org.sagebionetworks.repo.model.file.UploadType;
 import org.sagebionetworks.repo.model.jdo.NameValidation;
+import org.sagebionetworks.repo.model.project.ExternalGoogleCloudStorageLocationSetting;
+import org.sagebionetworks.repo.model.project.ExternalS3StorageLocationSetting;
+import org.sagebionetworks.repo.model.project.S3StorageLocationSetting;
 import org.sagebionetworks.repo.model.project.StorageLocationSetting;
 import org.sagebionetworks.repo.transactions.WriteTransaction;
 import org.sagebionetworks.schema.adapter.JSONEntity;
 import org.sagebionetworks.schema.adapter.JSONObjectAdapterException;
 import org.sagebionetworks.schema.adapter.org.json.EntityFactory;
+import org.sagebionetworks.upload.multipart.GoogleCloudStorageMultipartUploadDAO;
 import org.sagebionetworks.upload.multipart.S3MultipartUploadDAO;
 import org.sagebionetworks.util.ValidateArgument;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -50,6 +59,9 @@ public class MultipartManagerV2Impl implements MultipartManagerV2 {
 
 	@Autowired
 	S3MultipartUploadDAO s3multipartUploadDAO;
+
+	@Autowired
+	GoogleCloudStorageMultipartUploadDAO googleCloudStorageMultipartUploadDAO;
 
 	@Autowired
 	MultipartUploadDAO multipartUploadDAO;
@@ -124,25 +136,35 @@ public class MultipartManagerV2Impl implements MultipartManagerV2 {
 	 * @return
 	 */
 	private CompositeMultipartUploadStatus createNewMultipartUpload(
-			UserInfo user, MultipartUploadRequest request, String requestMD5Hex) {
+			UserInfo user, MultipartUploadRequest request, String requestMD5Hex) throws RuntimeException{
 		// This is the first time we have seen this request.
-		StorageLocationSetting locationSettings = getStorageLocationSettions(request
-				.getStorageLocationId());
+		StorageLocationSetting locationSettings = getStorageLocationSettings(request.getStorageLocationId());
 		// the bucket depends on the upload location used.
 		String bucket = MultipartUtils.getBucket(locationSettings);
 		// create a new key for this file.
 		String key = MultipartUtils.createNewKey(user.getId().toString(),
 				request.getFileName(), locationSettings);
-		String uploadToken = s3multipartUploadDAO.initiateMultipartUpload(
-				bucket, key, request);
 		String requestJson = createRequestJSON(request);
+
+		String uploadToken = "";
+		UploadType uploadType;
+		if (locationSettings instanceof S3StorageLocationSetting || locationSettings instanceof ExternalS3StorageLocationSetting) {
+			uploadType = UploadType.S3;
+			uploadToken = s3multipartUploadDAO.initiateMultipartUpload(
+					bucket, key, request);
+		} else if (locationSettings instanceof ExternalGoogleCloudStorageLocationSetting) {
+			uploadType = UploadType.GOOGLECLOUDSTORAGE;
+		} else {
+			throw new RuntimeException(); // TODO
+		}
+
 		// How many parts will be needed to upload this file?
 		int numberParts = PartUtils.calculateNumberOfParts(request.getFileSizeBytes(),
 				request.getPartSizeBytes());
 		// Start the upload
 		return multipartUploadDAO
 				.createUploadStatus(new CreateMultipartRequest(user.getId(),
-						requestMD5Hex, requestJson, uploadToken, bucket, key,
+						requestMD5Hex, requestJson, uploadToken, uploadType, bucket, key,
 						numberParts));
 	}
 
@@ -192,7 +214,7 @@ public class MultipartManagerV2Impl implements MultipartManagerV2 {
 		}
 	}
 
-	private StorageLocationSetting getStorageLocationSettions(Long storageId) {
+	private StorageLocationSetting getStorageLocationSettings(Long storageId) {
 		if (storageId == null) {
 			return null;
 		}
@@ -246,14 +268,22 @@ public class MultipartManagerV2Impl implements MultipartManagerV2 {
 
 		// validate the caller is the user that started this upload.
 		validateStartedBy(user, status);
-		List<PartPresignedUrl> partUrls = new LinkedList<PartPresignedUrl>();
+		List<PartPresignedUrl> partUrls = new LinkedList<>();
 		for (Long partNumberL : request.getPartNumbers()) {
 			ValidateArgument.required(partNumberL, "PartNumber cannot be null");
 			int partNumber = partNumberL.intValue();
 			validatePartNumber(partNumber, numberOfParts);
 			String partKey = createPartKey(status.getKey(), partNumber);
-			URL url = s3multipartUploadDAO.createPreSignedPutUrl(
-					status.getBucket(), partKey, request.getContentType());
+			URL url;
+			if (status.getUploadType().equals(UploadType.GOOGLECLOUDSTORAGE)) {
+				url = googleCloudStorageMultipartUploadDAO.createSignedPutUrl(
+						status.getBucket(), partKey, request.getContentType());
+			} else if (status.getUploadType().equals(UploadType.S3)) {
+				url = s3multipartUploadDAO.createPreSignedPutUrl(
+						status.getBucket(), partKey, request.getContentType());
+			} else {
+				throw new RuntimeException(); // TODO
+			}
 			PartPresignedUrl part = new PartPresignedUrl();
 			part.setPartNumber((long) partNumber);
 			part.setUploadPresignedUrl(url.toString());
@@ -302,7 +332,6 @@ public class MultipartManagerV2Impl implements MultipartManagerV2 {
 	 * Validate the caller started the given upload.
 	 * 
 	 * @param user
-	 * @param startedBy
 	 */
 	public static void validateStartedBy(UserInfo user,
 			CompositeMultipartUploadStatus composite) {
@@ -345,15 +374,29 @@ public class MultipartManagerV2Impl implements MultipartManagerV2 {
 		response.setPartNumber(new Long(partNumber));
 		response.setUploadId(uploadId);
 		try {
-			s3multipartUploadDAO.addPart(new AddPartRequest(composite
-					.getUploadToken(), composite.getBucket(), composite
-					.getKey(), partKey, partMD5Hex, partNumber));
+			if (composite.getUploadType().equals(UploadType.S3)) {
+				s3multipartUploadDAO.addPart(new AddPartRequest(composite
+						.getUploadToken(), composite.getBucket(), composite
+						.getKey(), partKey, partMD5Hex, partNumber));
+			} else if (composite.getUploadType().equals(UploadType.GOOGLECLOUDSTORAGE)) {
+				googleCloudStorageMultipartUploadDAO.addPart(new AddPartRequest(composite.getUploadToken(),
+						composite.getBucket(), composite.getKey(), partKey, partMD5Hex, partNumber));
+			} else {
+				throw new IllegalArgumentException("Unknown multipart upload type");
+			}
 			// added the part successfully.
 			multipartUploadDAO
 					.addPartToUpload(uploadId, partNumber, partMD5Hex);
 			response.setAddPartState(AddPartState.ADD_SUCCESS);
-			// after a part is added we can delete the part file
-			s3multipartUploadDAO.deleteObject(composite.getBucket(), partKey);
+			// after a part is added to S3 we can delete the part file
+			if (composite.getUploadType().equals(UploadType.S3)) {
+				s3multipartUploadDAO.deleteObject(composite.getBucket(), partKey);
+			}
+
+			// we cannot do the same for Google Cloud, but we can rename the part to its MD5 so we can find it later
+			if (composite.getUploadType().equals(UploadType.GOOGLECLOUDSTORAGE)) {
+				googleCloudStorageMultipartUploadDAO.renameObject(composite.getBucket(), partKey, partMD5Hex);
+			}
 		} catch (Exception e) {
 			response.setErrorMessage(e.getMessage());
 			response.setAddPartState(AddPartState.ADD_FAILED);
@@ -390,16 +433,22 @@ public class MultipartManagerV2Impl implements MultipartManagerV2 {
 		request.setBucket(composite.getBucket());
 		request.setKey(composite.getKey());
 		request.setUploadToken(composite.getUploadToken());
-		// This will create the object in S3
-		long fileContentSize = s3multipartUploadDAO
-				.completeMultipartUpload(request);
-
+		long fileContentSize;
+		if (composite.getUploadType().equals(UploadType.S3)) {
+			// This will create the object in S3
+			fileContentSize = s3multipartUploadDAO.completeMultipartUpload(request);
+		} else if (composite.getUploadType().equals(UploadType.GOOGLECLOUDSTORAGE)) {
+			fileContentSize = googleCloudStorageMultipartUploadDAO.completeMultipartUpload(request, addedParts);
+		} else {
+			throw new RuntimeException(); // TODO:
+		}
 		// Get the original request
 		MultipartUploadRequest originalRequest = getRequestForUpload(uploadId);
 		// create a file handle to represent this file.
-		S3FileHandle resultFileHandle = createFileHandle(fileContentSize, composite, originalRequest);
+
+		String resultFileHandleId = createFileHandle(fileContentSize, composite, originalRequest).getId();
 		// complete the upload.
-		composite = multipartUploadDAO.setUploadComplete(uploadId, resultFileHandle.getId());
+		composite = multipartUploadDAO.setUploadComplete(uploadId, resultFileHandleId);
 		return prepareCompleteStatus(composite);
 	}
 
@@ -468,26 +517,54 @@ public class MultipartManagerV2Impl implements MultipartManagerV2 {
 	 */
 	@WriteTransaction
 	@Override
-	public S3FileHandle createFileHandle(long fileSize, CompositeMultipartUploadStatus composite, MultipartUploadRequest request){
-		// Convert all of the data to a file handle.
-		S3FileHandle fileHandle = new S3FileHandle();
-		fileHandle.setFileName(request.getFileName());
-		fileHandle.setContentType(request.getContentType());
-		fileHandle.setBucketName(composite.getBucket());
-		fileHandle.setKey(composite.getKey());
-		fileHandle.setCreatedBy(composite.getMultipartUploadStatus().getStartedBy());
-		fileHandle.setCreatedOn(new Date(System.currentTimeMillis()));
-		fileHandle.setEtag(UUID.randomUUID().toString());
-		fileHandle.setContentMd5(request.getContentMD5Hex());
-		fileHandle.setStorageLocationId(request.getStorageLocationId());
-		fileHandle.setContentSize(fileSize);
-		// By default a preview should be created.	
-		fileHandle.setId(idGenerator.generateNewId(IdType.FILE_IDS).toString());
-		if(request.getGeneratePreview() != null && !request.getGeneratePreview()){
-			fileHandle.setPreviewId(fileHandle.getId());
+	public S3FileHandleInterface createFileHandle(long fileSize, CompositeMultipartUploadStatus composite, MultipartUploadRequest request){
+		// TODO: cleanup
+		switch (composite.getUploadType()) {
+			case S3:
+				S3FileHandle s3FileHandle = new S3FileHandle();
+				s3FileHandle.setFileName(request.getFileName());
+				s3FileHandle.setContentType(request.getContentType());
+				s3FileHandle.setBucketName(composite.getBucket());
+				s3FileHandle.setKey(composite.getKey());
+				s3FileHandle.setCreatedBy(composite.getMultipartUploadStatus().getStartedBy());
+				s3FileHandle.setCreatedOn(new Date(System.currentTimeMillis()));
+				s3FileHandle.setEtag(UUID.randomUUID().toString());
+				s3FileHandle.setContentMd5(request.getContentMD5Hex());
+				s3FileHandle.setStorageLocationId(request.getStorageLocationId());
+				s3FileHandle.setContentSize(fileSize);
+				// By default a preview should be created.
+				s3FileHandle.setId(idGenerator.generateNewId(IdType.FILE_IDS).toString());
+				if(request.getGeneratePreview() != null && !request.getGeneratePreview()){
+					s3FileHandle.setPreviewId(s3FileHandle.getId());
+				}
+				return (S3FileHandleInterface) fileHandleDao.createFile(s3FileHandle);
+			case GOOGLECLOUDSTORAGE:
+				GoogleCloudFileHandle googleCloudFileHandle = new GoogleCloudFileHandle();
+				googleCloudFileHandle.setFileName(request.getFileName());
+				googleCloudFileHandle.setContentType(request.getContentType());
+				googleCloudFileHandle.setBucketName(composite.getBucket());
+				googleCloudFileHandle.setKey(composite.getKey());
+				googleCloudFileHandle.setCreatedBy(composite.getMultipartUploadStatus().getStartedBy());
+				googleCloudFileHandle.setCreatedOn(new Date(System.currentTimeMillis()));
+				googleCloudFileHandle.setEtag(UUID.randomUUID().toString());
+				googleCloudFileHandle.setContentMd5(request.getContentMD5Hex());
+				googleCloudFileHandle.setStorageLocationId(request.getStorageLocationId());
+				googleCloudFileHandle.setContentSize(fileSize);
+				// By default a preview should be created.
+				googleCloudFileHandle.setId(idGenerator.generateNewId(IdType.FILE_IDS).toString());
+				if(request.getGeneratePreview() != null && !request.getGeneratePreview()){
+					googleCloudFileHandle.setPreviewId(googleCloudFileHandle.getId());
+				}
+				return (S3FileHandleInterface) fileHandleDao.createFile(googleCloudFileHandle);
+			case SFTP:
+			case HTTPS:
+			case PROXYLOCAL:
+			default:
+				throw new IllegalArgumentException("Multipart upload only supports S3 and Google Cloud");
 		}
+		// TODO: platform independence
+		// Convert all of the data to a file handle.
 		// dao creates the files handle.
-		return (S3FileHandle) fileHandleDao.createFile(fileHandle);
 	}
 
 	@Override
